fix: carry a motif position as a double, not a float
compute_position_result() built the reported position as `float(index) + 1.0f`
and every function returning it was declared float. binary32 has a 24-bit
significand, so consecutive integers stop being representable past 2^24: a
unique match planted at 0-based offset 16,777,217 of an interval on mm10 chr1
came back as 16,777,216 where 16,777,218 is right. A whole-chromosome iterator
reaches that range on any mammalian assembly.

The scan-limit fix in the previous commit is what made this reachable on the
forward strand at all; until then the 1e6 cap stopped the scan long before a
position could get that large, and the two defects hid each other.

Only the RETURN types are widened - the accumulators stay float - so every
score is bit-identical to before and only positions change. Verified: the
example-genome score checksums for pwm, pwm.max and pwm.count match the
previous build exactly, and the full 199-file suite is identical to
origin/master on every file (3 failed / 207 errored on both, all in
genome-build and multitasking helpers unrelated to this code).

+++ src/KmerCounter.cpp
@@ -34,7 +34,7 @@ KmerCounter::KmerCounter(const std::string &kmer, GenomeSeqFetch* shared_seqfetc
                    [](unsigned char c) { return std::toupper(c); });
 }
 
-float KmerCounter::score_interval(const GInterval &interval, const GenomeChromKey &chromkey)
+double KmerCounter::score_interval(const GInterval &interval, const GenomeChromKey &chromkey)
 {
     // Safety check - shouldn't happen due to constructor validation
     if (m_kmer.empty())
